+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to `LaravelHelpers` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - YYYY-MM-DD
+## 2018-03-30
 
 ### Added
-- Nothing
+- Added option to set thumbnail size
 
 ### Deprecated
 - Nothing